// src/router/types/route-names.ts
export const enum RouteNames {
  // SHARED
  HOME = 'HOME',
  PROJECT = 'PROJECT',
  PROFILE = 'PROFILE',
  FAQ = 'FAQ',
  CONTACTS = 'CONTACTS',
  // project
  PROJECT_DETAILS = 'PROJECT_DETAILS',
  PROJECT_PARTICIPATIONS = 'PROJECT_PARTICIPATIONS',
  PROJECT_PARTICIPANTS = 'PROJECT_PARTICIPANTS',
  NOT_FOUND = 'NOT_FOUND',

  // USER
  USER_INFO = 'USER_INFO',
  USER_PROJECTS = 'USER_PROJECTS',
  // candidate
  CANDIDATE_PARTICIPATIONS = 'CANDIDATE_PARTICIPATIONS',
  // supervisor
  SUPERVISOR_PROJECT_PROPOSAL_CREATE = 'SUPERVISOR_PROJECT_PROPOSAL_CREATE',
  SUPERVISOR_PROJECT_PROPOSAL_RESULT = 'SUPERVISOR_PROJECT_PROPOSAL_RESULT',
  SUPERVISOR_PROJECT_PROPOSALS = 'SUPERVISOR_PROJECT_PROPOSALS',
  // institute director
  INST_DIRECTOR_PROJECTS = 'INST_DIRECTOR_PROJECTS',
  INST_DIRECTOR_PROJECTS_ALL = 'INST_DIRECTOR_PROJECTS_ALL',
  INST_DIRECTOR_PROJECTS_ACTIVE = 'INST_DIRECTOR_PROJECTS_ACTIVE',
  INST_DIRECTOR_PROJECT_PROPOSALS = 'INST_DIRECTOR_PROJECT_PROPOSALS',
  INST_DIRECTOR_PROJECT_PROPOSALS_NEW = 'INST_DIRECTOR_PROJECT_PROPOSALS_NEW',
  INST_DIRECTOR_PROJECT_PROPOSALS_REJECTED = 'INST_DIRECTOR_PROJECT_PROPOSALS_REJECTED',
  INST_DIRECTOR_PROJECT_PROPOSALS_APPROVED = 'INST_DIRECTOR_PROJECT_PROPOSALS_APPROVED',
  INST_DIRECTOR_PROJECT_PROPOSALS_APPROVED_SPRING = 'INST_DIRECTOR_PROJECT_PROPOSALS_APPROVED_SPRING',
  INST_DIRECTOR_PROJECT_PROPOSALS_APPROVED_AUTUMN = 'INST_DIRECTOR_PROJECT_PROPOSALS_APPROVED_AUTUMN',
}